[Enhancement] Cria a logica para cadastrar receitas

+++ src/main/kotlin/br/com/gabi/ControleFinanceiro/dto/NovaReceitaDto.kt
@@ -1,11 +1,10 @@
 package br.com.gabi.ControleFinanceiro.dto
 
-import java.math.BigDecimal
 import java.time.LocalDateTime
 import javax.validation.constraints.NotEmpty
 
 data class NovaReceitaDto(
     @NotEmpty val descricao: String,
-    @NotEmpty val valor: BigDecimal,
+    @NotEmpty val valor: Double,
     @NotEmpty val data: LocalDateTime = LocalDateTime.now()
 )
